Move registry api definitions under registry package
