+++ registry/api/v2/errors.go
