+++ registry/api/v2/errors_test.go
