+++ registry/api/v2/routes.go
