+++ registry/api/v2/routes_test.go
