+++ registry/handlers/app.go
@@ -7,7 +7,7 @@ import (
 	"os"
 
 	"code.google.com/p/go-uuid/uuid"
-	"github.com/docker/distribution/api/v2"
+	"github.com/docker/distribution/registry/api/v2"
 	"github.com/docker/distribution/auth"
 	"github.com/docker/distribution/configuration"
 	ctxu "github.com/docker/distribution/context"
